% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_crosswalk.R
\name{common_name_to_plant_code}
\alias{common_name_to_plant_code}
\title{Convert biomass common names to OSM PlantCodes}
\usage{
common_name_to_plant_code(common_name)
}
\arguments{
\item{common_name}{character vector of lowercase common names.}
}
\value{
character vector of OSM \code{PlantCodes} strings, same length as
  \code{common_name}. \code{NA} is returned with a warning for
  unrecognised names.
}
\description{
Translates one or more lowercase common species names (as used in
\code{\link{biomass_tree}} and \code{\link{biomass_coefs}}) to OSM
\code{PlantCodes} strings.
}
\seealso{
\code{\link{plant_codes}}, \code{\link{bc_species_to_osm}}
}
