% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_crosswalk.R
\name{plant_code_to_common_name}
\alias{plant_code_to_common_name}
\title{Convert OSM PlantCodes to biomass common names}
\usage{
plant_code_to_common_name(plant_code)
}
\arguments{
\item{plant_code}{character vector of OSM \code{PlantCodes} strings.}
}
\value{
character vector of common names, same length as
  \code{plant_code}. \code{NA} is returned (with a warning) for
  unrecognised codes or species with no biomass equation.
}
\description{
Translates one or more OSM \code{PlantCodes} strings to the lowercase
common names used in \code{\link{biomass_tree}}.  Returns \code{NA} for
species that do not have biomass equations in this package.
}
\seealso{
\code{\link{plant_codes}}, \code{\link{bc_species_to_osm}}
}
