% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_crosswalk.R
\name{plant_code_to_sp0}
\alias{plant_code_to_sp0}
\title{Convert OSM PlantCodes to group-level species codes (SP0)}
\usage{
plant_code_to_sp0(plant_code)
}
\arguments{
\item{plant_code}{character vector of OSM \code{PlantCodes} strings.}
}
\value{
character vector of group-level species codes (SP0), same length as
  \code{plant_code}.  \code{NA} is returned (with a warning) for
  unrecognised codes or species with no BC taper equation.
}
\description{
Translates one or more OSM \code{PlantCodes} strings to group-level
species codes (referred to as SP0 in BC inventory) used in
\code{\link{tree_volume}}.  Returns \code{NA} for species
that do not have a BC taper equation in \code{\link{taper_coefs_kbec}}.
}
\seealso{
\code{\link{plant_codes}}, \code{\link{bc_species_to_osm}}
}
